Refactored message parsing
